# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DocSecur API",
        "description": "API REST pour la plateforme de gestion médicale DocSecur. Authentification via Bearer Token (Laravel Passport).",
        "contact": {
            "email": "support@docsecur.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
        "/api/admin/campagnes": {
            "get": {
                "tags": [
                    "Admin - Campagnes"
                ],
                "summary": "Lister les campagnes de santé",
                "description": "Retourne la liste paginée des campagnes, triées par date de début décroissante. Rôle requis : administrateur.",
                "operationId": "367726e7453bddda78d9cd6b68d6dc40",
                "responses": {
                    "200": {
                        "description": "Liste des campagnes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Campagnes"
                ],
                "summary": "Créer une campagne de santé",
                "description": "Crée une nouvelle campagne (prévention, vaccination ou sensibilisation). Rôle requis : administrateur.",
                "operationId": "d7a67a39f6547fca030112d126276ba4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "date_debut",
                                    "type"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Campagne de vaccination contre la polio"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Campagne nationale de vaccination"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-01"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-30"
                                    },
                                    "cible": {
                                        "type": "string",
                                        "example": "Enfants de 0 à 5 ans"
                                    },
                                    "region": {
                                        "type": "string",
                                        "example": "Dakar"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "prevention",
                                            "vaccination",
                                            "sensibilisation"
                                        ],
                                        "example": "vaccination"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campagne créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Campagne créée"
                                        },
                                        "campagne": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/campagnes/{id}": {
            "get": {
                "tags": [
                    "Admin - Campagnes"
                ],
                "summary": "Détails d'une campagne",
                "operationId": "e95d81111a80acc1fd128f20b125e9e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la campagne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campagne non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Campagnes"
                ],
                "summary": "Modifier une campagne",
                "operationId": "16a967d4c85a4325158b380dbcd2f8bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "cible": {
                                        "type": "string"
                                    },
                                    "region": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "prevention",
                                            "vaccination",
                                            "sensibilisation"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campagne modifiée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Campagne modifiée"
                                        },
                                        "campagne": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campagne non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Campagnes"
                ],
                "summary": "Supprimer une campagne",
                "operationId": "a26e3fec2198b870113544a70655daf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campagne supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Campagne supprimée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campagne non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}